Love load_results_manually2 to test_enkf_main
